testgl/testgl/requirements.txt
pygame
PyOpenGL
aenum
